[SYCL] Fix test configurations when L0 and OpenCL are disabled (#8247)
This commit changes the configuration of in-tree SYCL LIT tests to
include features for when the L0 and OpenCL plugins are enabled. This
allows the interop backend tests to be split into backend-dependent
tests to be run only when their backends are enabled. Additionally, this
commit changes the fsycl-host-only LIT configuration replacement to
ignore unset include paths, avoiding empty "-isystem" options when
related backends are disabled.
