+++ sycl/test/lit.site.cfg.py.in
@@ -28,6 +28,8 @@ config.extra_environment = lit_config.params.get("extra_environment", "@LIT_EXTR
 config.cuda_be = '@SYCL_BUILD_PI_CUDA@'
 config.esimd_emulator_be = '@SYCL_BUILD_PI_ESIMD_EMULATOR@'
 config.hip_be = '@SYCL_BUILD_PI_HIP@'
+config.opencl_be = '@SYCL_BUILD_PI_OPENCL@'
+config.level_zero_be = '@SYCL_BUILD_PI_LEVEL_ZERO@'
 
 import lit.llvm
 lit.llvm.initialize(lit_config, config)
